refactor: 기존 /main 의 page.tsx 컴포넌트 compound pattern을 사용하여 리팩토링 (ProgramWrapper.tsx)
기존에 page.tsx 에서 파라미터를 가져와  리다이렉트 및 상태 변경 로직이 모두 셖여있어, 재사용하기 어려운 상태였다.
admin 에서도 동일한 ui및 동작을 보여주어야 하므로, 이를 재사용 가능하도록 compound pattern을 사용하여 수정
