+++ FE/src/components/main/Program.tsx
@@ -1,0 +1,14 @@
+"use client";
+import ProgramWrapper from "./Program.compound";
+
+const Program = () => {
+  return (
+    <ProgramWrapper>
+      <ProgramWrapper.CategoryTab />
+      <ProgramWrapper.StatusTab />
+      <ProgramWrapper.Content />
+    </ProgramWrapper>
+  );
+};
+
+export default Program;
